Resolve "add boxes"

Closes #8

See merge request alexkochetov/learningcards!9

+++ src/main/java/dev/alex/cards/model/Box.java
@@ -17,7 +17,7 @@ import javax.persistence.OneToMany;
 public class Box {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
     private String name;
     @Column(name = "repeat_in")
     private int amountDays;
@@ -60,12 +60,12 @@ public class Box {
     public int getElementsCounter() {
         return cards.size();
     }
-    
+
     public List<Card> getCardsByTheme(Theme theme) {
-    	List<Card> cards = getCards();
-    	return cards.stream().filter(card -> card.getTheme() == theme).collect(Collectors.toList());
+        List<Card> cards = getCards();
+        return cards.stream().filter(card -> card.getTheme() == theme).collect(Collectors.toList());
     }
-    
+
     public List<Card> getCards() {
         return cards;
     }
